refactor service interface삭제
