+++ src/test/java/com/programmers/service/FoodServiceTest.java
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @SpringBootTest
-class FoodServiceImplTest {
+class FoodServiceTest {
     @Autowired
     FoodRepository foodRepository;
